Improvement CharacterDetailView style and minor refactorin
