+++ RickAndMorty_API/View/Scenes/CharactersView/ViewStates/CharactersListView.swift
